split core streamdaq methods to node functions
